+++ lib/easy_stepper.dart
@@ -4,11 +4,13 @@ import 'dart:io';
 import 'dart:math';
 
 import 'package:easy_stepper/src/core/custom_scroll_behavior.dart';
+import 'package:easy_stepper/src/core/line_style.dart';
 import 'package:easy_stepper/src/easy_step.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 export 'package:easy_stepper/src/easy_step.dart';
+export 'package:easy_stepper/src/core/line_style.dart';
 
 class EasyStepper extends StatefulWidget {
   /// Each Step defines a step icon and title. Hence, total number of icons determines the total number of steps.
@@ -68,19 +70,6 @@ class EasyStepper extends StatefulWidget {
   /// The type of the step border  when it is finished [normal, dotted].
   final BorderType? finishedStepBorderType;
 
-  /// The color of the line that separates the steps.
-  /// If null, [Theme.of(context).colorScheme.primary] will be used.
-  final Color? defaultLineColor;
-
-  /// The color of the line that branches out from unreached steps.
-  final Color? unreachedLineColor;
-
-  /// The color of the line that branches out from active step.
-  final Color? activeLineColor;
-
-  /// The color of the line that branches out from finished steps.
-  final Color? finishedLineColor;
-
   /// The radius of a step.
   final double stepRadius;
 
@@ -105,38 +94,17 @@ class EasyStepper extends StatefulWidget {
   /// Specifies the alignment of IconStepper widget.
   final AlignmentGeometry alignment;
 
-  /// The length of the line that separates the steps.
-  final double lineLength;
-
-  /// The radius of individual dot within the line that separates the steps, if `lineType` set to `normal` this property is used to change the
-  final double? lineDotRadius;
-
-  /// The thickness of the line that separates the steps.
-  final double lineThickness;
-
-  /// The space between individual dot within the line that separates the steps.
-  final double lineSpace;
-
-  /// The type of the line [normal, dotted].
-  final LineType lineType;
-
   /// The amount of padding around every step.
   final double internalPadding;
 
   /// The amount of padding around the stepper.
   final EdgeInsetsGeometry padding;
 
-  /// The animation effect to show when a step is reached.
-  final Curve stepReachedAnimationEffect;
-
   /// The curve of the animation to show when a step is reached.
-  final Curve? stepAnimationCurve;
+  final Curve stepAnimationCurve;
 
   /// The duration of the animation effect to show when a step is reached.
-  final Duration stepReachedAnimationDuration;
-
-  /// The duration of the animation effect to show when a step is reached.
-  final Duration? stepAnimationDuration;
+  final Duration stepAnimationDuration;
 
   ///The thickness of the step border.
   final double borderThickness;
@@ -172,13 +140,15 @@ class EasyStepper extends StatefulWidget {
   /// Whether the stepper take the full width of the screen or not, this property work when `disableScroll = true`. default `True`.
   final bool fitWidth;
 
+  //All Styles to customize line between steps
+  final LineStyle? lineStyle;
+
   const EasyStepper({
     Key? key,
     required this.activeStep,
     required this.steps,
     this.reachedSteps,
     this.maxReachedStep,
-    this.lineType = LineType.dotted,
     this.enableStepTapping = true,
     this.direction = Axis.horizontal,
     this.onStepReached,
@@ -194,34 +164,18 @@ class EasyStepper extends StatefulWidget {
     this.finishedStepBackgroundColor,
     this.finishedStepBorderColor,
     this.finishedStepIconColor,
-    this.defaultLineColor,
-    this.unreachedLineColor,
-    this.activeLineColor,
-    this.finishedLineColor,
     this.stepRadius = 30,
     this.steppingEnabled = true,
     this.disableScroll = false,
     this.showTitle = true,
     this.alignment = Alignment.center,
-    this.lineLength = 40,
     this.fitWidth = true,
     this.showScrollbar = true,
-    @Deprecated(
-        "use 'lineThickness' instead, This feature was deprecated after v0.5.1")
-    this.lineDotRadius,
-    this.lineThickness = 1,
-    this.lineSpace = 5,
     this.padding =
         const EdgeInsetsDirectional.symmetric(horizontal: 10, vertical: 10),
     this.internalPadding = 8,
-    @Deprecated(
-        "use 'stepAnimationCurve' instead, This feature was deprecated after v0.1.4+1")
-    this.stepReachedAnimationEffect = Curves.linear,
-    this.stepAnimationCurve,
-    @Deprecated(
-        "use 'stepAnimationDuration' instead, This feature was deprecated after v0.1.4+1")
-    this.stepReachedAnimationDuration = const Duration(seconds: 1),
-    this.stepAnimationDuration,
+    this.stepAnimationCurve = Curves.linear,
+    this.stepAnimationDuration = const Duration(seconds: 1),
     this.borderThickness = 0.8,
     this.loadingAnimation,
     this.stepShape = StepShape.circle,
@@ -234,7 +188,10 @@ class EasyStepper extends StatefulWidget {
     this.showStepBorder = true,
     this.showLoadingAnimation = true,
     this.textDirection = TextDirection.ltr,
-  }) : assert(maxReachedStep == null || reachedSteps == null, 'only "maxReachedStep" or "reachedSteps" allowed'), super(key: key);
+    this.lineStyle,
+  })  : assert(maxReachedStep == null || reachedSteps == null,
+            'only "maxReachedStep" or "reachedSteps" allowed'),
+        super(key: key);
 
   @override
   State<EasyStepper> createState() => _EasyStepperState();
@@ -243,9 +200,11 @@ class EasyStepper extends StatefulWidget {
 class _EasyStepperState extends State<EasyStepper> {
   ScrollController? _scrollController;
   late int _selectedIndex;
+  late LineStyle lineStyle;
 
   @override
   void initState() {
+    lineStyle = widget.lineStyle ?? const LineStyle();
     _selectedIndex = widget.activeStep;
     _scrollController = ScrollController();
     super.initState();
@@ -274,10 +233,9 @@ class _EasyStepperState extends State<EasyStepper> {
         i *
             ((widget.stepRadius * 2) +
                 widget.internalPadding +
-                widget.lineLength),
-        duration:
-            widget.stepAnimationDuration ?? widget.stepReachedAnimationDuration,
-        curve: widget.stepAnimationCurve ?? widget.stepReachedAnimationEffect,
+                lineStyle.lineLength),
+        duration: widget.stepAnimationDuration,
+        curve: widget.stepAnimationCurve,
       );
 
       if (_selectedIndex == i) break;
@@ -286,6 +244,7 @@ class _EasyStepperState extends State<EasyStepper> {
 
   @override
   Widget build(BuildContext context) {
+    lineStyle = widget.lineStyle ?? const LineStyle();
     // Controls scrolling behavior.
     if (!widget.disableScroll) {
       WidgetsBinding.instance.addPostFrameCallback(_afterLayout);
@@ -363,13 +322,13 @@ class _EasyStepperState extends State<EasyStepper> {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
                 _buildStep(index),
-                _buildDottedLine(index, Axis.horizontal),
+                _buildLine(index, Axis.horizontal),
               ],
             )
           : Column(
               children: <Widget>[
                 _buildStep(index),
-                _buildDottedLine(index, Axis.vertical),
+                _buildLine(index, Axis.vertical),
               ],
             );
     });
@@ -387,10 +346,11 @@ class _EasyStepperState extends State<EasyStepper> {
           ? index < widget.activeStep && widget.reachedSteps!.contains(index)
           : index < widget.activeStep,
       isUnreached: index > widget.activeStep,
-      isAlreadyReached: widget.reachedSteps != null ? widget.reachedSteps!.contains(index) :
-      widget.maxReachedStep != null
-          ? index <= widget.maxReachedStep!
-          : false,
+      isAlreadyReached: widget.reachedSteps != null
+          ? widget.reachedSteps!.contains(index)
+          : widget.maxReachedStep != null
+              ? index <= widget.maxReachedStep!
+              : false,
       activeStepBackgroundColor: widget.activeStepBackgroundColor,
       activeStepBorderColor: widget.activeStepBorderColor,
       activeTextColor: widget.activeStepTextColor,
@@ -413,7 +373,7 @@ class _EasyStepperState extends State<EasyStepper> {
       showStepBorder: widget.showStepBorder,
       showLoadingAnimation: widget.showLoadingAnimation,
       textDirection: widget.textDirection,
-      lineLength: widget.lineLength,
+      lineLength: lineStyle.lineLength,
       enabled: widget.steps[index].enabled,
       direction: widget.direction,
       onStepSelected: widget.enableStepTapping
@@ -449,46 +409,40 @@ class _EasyStepperState extends State<EasyStepper> {
     Color? preferredColor;
     if (index == widget.activeStep) {
       //Active Step
-      preferredColor = widget.activeLineColor;
+      preferredColor = lineStyle.activeLineColor;
     } else if (index > widget.activeStep) {
       //Unreached Step
-      preferredColor = widget.unreachedLineColor;
+      preferredColor = lineStyle.unreachedLineColor;
     } else if (index < widget.activeStep) {
       //Finished Step
-      preferredColor = widget.finishedLineColor;
+      preferredColor = lineStyle.finishedLineColor;
     }
 
     return preferredColor ??
-        widget.defaultLineColor ??
+        lineStyle.defaultLineColor ??
         Theme.of(context).colorScheme.primary;
   }
 
-  Widget _buildDottedLine(int index, Axis axis) {
+  Widget _buildLine(int index, Axis axis) {
     return index < widget.steps.length - 1
         ? Column(
             children: [
               Padding(
                 padding: EdgeInsets.only(
                   top: axis == Axis.horizontal
-                      ? (widget.stepRadius -
-                          (widget.lineDotRadius ?? widget.lineThickness))
+                      ? (widget.stepRadius - lineStyle.lineThickness)
                       : 0,
                   bottom: axis == Axis.vertical && widget.showTitle ? 10 : 0,
                 ),
-                child: EasyLine(
-                  length: widget.lineLength,
-                  color: _getLineColor(index),
-                  thickness: widget.lineDotRadius ?? widget.lineThickness,
-                  spacing: widget.lineSpace,
-                  axis: axis,
-                  lineType: widget.lineType,
-                ),
+                child: lineStyle.progress != null && index == widget.activeStep
+                    ? _buildProgressLine(index, axis)
+                    : _buildBaseLine(index, axis),
               ),
               if (axis == Axis.horizontal &&
                   widget.steps[index].lineText != null) ...[
                 const SizedBox(height: 5),
                 SizedBox(
-                  width: widget.lineLength,
+                  width: lineStyle.lineLength,
                   child: widget.steps[index].customLineWidget ??
                       Text(
                         widget.steps[index].lineText!,
@@ -501,5 +455,48 @@ class _EasyStepperState extends State<EasyStepper> {
             ],
           )
         : const Offstage();
+  }
+
+  Stack _buildProgressLine(int index, Axis axis) {
+    return Stack(
+      children: [
+        _buildBaseLine(index, axis),
+        PositionedDirectional(
+          start: 0,
+          top: 0,
+          bottom: 0,
+          child: AnimatedContainer(
+            duration: const Duration(milliseconds: 400),
+            height: lineStyle.lineThickness,
+            width: lineStyle.lineLength * lineStyle.progress!,
+            child: EasyLine(
+              length: lineStyle.lineLength * lineStyle.progress!,
+              color: lineStyle.progressColor ??
+                  lineStyle.activeLineColor ??
+                  Theme.of(context).colorScheme.primary,
+              thickness: lineStyle.lineThickness,
+              spacing: lineStyle.lineSpace,
+              axis: axis,
+              lineType: lineStyle.lineType,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  EasyLine _buildBaseLine(int index, Axis axis) {
+    return EasyLine(
+      length: lineStyle.lineLength,
+      color: _getLineColor(index),
+      thickness: lineStyle.lineThickness,
+      spacing: lineStyle.lineSpace,
+      width: lineStyle.lineWidth,
+      axis: axis,
+      lineType:
+          index > widget.activeStep - 1 && lineStyle.unreachedLineType != null
+              ? lineStyle.unreachedLineType!
+              : lineStyle.lineType,
+    );
   }
 }