+++ lib/src/core/easy_line.dart
@@ -4,6 +4,9 @@ class EasyLine extends StatelessWidget {
   /// Width of the dotted line.
   final double length;
 
+  /// Width of the dashes.
+  final double width;
+
   /// Color of the dotted line.
   final Color color;
 
@@ -23,8 +26,9 @@ class EasyLine extends StatelessWidget {
     key,
     this.length = 50.0,
     this.color = Colors.grey,
-    this.thickness = 1.5,
+    this.thickness = 3,
     this.spacing = 3.0,
+    this.width = 2.0,
     this.lineType = LineType.dotted,
     this.axis = Axis.horizontal,
   }) : super(key: key);
@@ -33,17 +37,20 @@ class EasyLine extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       // If this is not applied, top half of the dot gets offscreen, hence, hidden.
-      margin: EdgeInsets.only(top: lineType == LineType.dotted ? thickness : 0),
+      margin: EdgeInsets.only(
+          top: lineType == LineType.dotted || lineType == LineType.dashed
+              ? thickness
+              : 0),
       width: axis == Axis.horizontal
           ? length
-          : lineType == LineType.normal
-              ? thickness * 2
-              : 0,
+          : lineType == LineType.dotted
+              ? 0
+              : thickness,
       height: axis == Axis.vertical
           ? length
-          : lineType == LineType.normal
-              ? thickness * 2
-              : 0,
+          : lineType == LineType.dotted
+              ? 0
+              : thickness,
       decoration: lineType == LineType.normal
           ? BoxDecoration(
               color: color,
@@ -56,11 +63,29 @@ class EasyLine extends StatelessWidget {
                 brush: Paint()..color = color,
                 length: length,
                 dotRadius: thickness,
-                spacing: lineType == LineType.normal ? 0.0 : spacing,
+                spacing: spacing,
                 axis: axis,
               ),
             )
-          : null,
+          : lineType == LineType.dashed
+              ? Row(
+                  children: List.generate(
+                    length ~/ ((width + spacing) / 2),
+                    (index) => Expanded(
+                      child: index % 2 == 0
+                          ? Container(
+                              color: color,
+                              height: thickness,
+                              width: width,
+                            )
+                          : SizedBox(
+                              width: spacing,
+                              height: thickness,
+                            ),
+                    ),
+                  ),
+                )
+              : null,
     );
   }
 }
@@ -122,4 +147,7 @@ enum LineType {
 
   /// Dotted line with space between dots.
   dotted,
+
+  /// Dashed line with space between dashes.
+  dashed,
 }